+++ components/org.wso2.openbanking.berlin.dynamic.client.creator/pom.xml
@@ -17,7 +17,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>org.wso2.openbanking.berlin.executor</artifactId>
+    <artifactId>org.wso2.openbanking.berlin.dynamic.client.creator</artifactId>
     <packaging>bundle</packaging>
     <name>WSO2 Open Banking Berlin Toolkit - Dynamic Client Creation Executor</name>
     <dependencies>
