+++ components/pom.xml
@@ -42,6 +42,6 @@
        <module>org.wso2.openbanking.berlin.keymanager</module>
         <module>org.wso2.openbanking.berlin.consent.extensions</module>
         <module>org.wso2.openbanking.berlin.demo.backend</module>
-        <module>org.wso2.openbanking.berlin.executor</module>
+        <module>org.wso2.openbanking.berlin.dynamic.client.creator</module>
     </modules>
 </project>
